requirements.txt
deap
numpy
pyyaml
torch
botorch
gpytorch
pandas
pydantic
ipywidgets
tqdm
